refactor training and validation functions into classes
