+++ model/model_main.py
@@ -2,18 +2,16 @@ import os
 import os.path as op
 import tensorflow as tf
 import numpy as np
-import time
 
 import settings
 from config import opts
 from tfrecords.tfrecord_reader import TfrecordGenerator
 import utils.util_funcs as uf
 from utils.util_class import TrainException
-from model.loss_and_metric.loss_factory import loss_factory
-from model.loss_and_metric.metric import compute_metric_pose
 from model.build_model.model_factory import ModelFactory
 from model.optimizers import optimizer_factory
 import model.logger as log
+from model.train_model import ModelTrainerGraph, ModelValidaterGraph
 
 
 def train_by_user_interaction():
@@ -49,25 +47,27 @@ def train():
     # TODO WARNING! using "test" split for training dataset is just to check training process
     dataset_train, train_steps = get_dataset(opts.DATASET, "test", True)
     dataset_val, val_steps = get_dataset(opts.DATASET, "test", False)
+    trainer_graph = ModelTrainerGraph()
+    validater_graph = ModelValidaterGraph()
     optimizer = optimizer_factory("adam_constant", opts.LEARNING_RATE, initial_epoch)
 
     print(f"\n\n========== START TRAINING ON {opts.CKPT_NAME} ==========")
     for epoch in range(initial_epoch, opts.EPOCHS):
         print(f"========== Start epoch: {epoch}/{opts.EPOCHS} ==========")
 
-        result_train, depth_train = train_an_epoch_graph(model, dataset_train, optimizer, train_steps)
-        print(f"\n[Train Epoch MEAN], result: loss={result_train[0]:1.4f}, "
+        result_train, depth_train = trainer_graph.run_an_epoch(model, dataset_train, train_steps, optimizer)
+        print(f"[Train Epoch MEAN], result: loss={result_train[0]:1.4f}, "
               f"trj_err={result_train[1]:1.4f}, rot_err={result_train[2]:1.4f}")
 
-        result_val, depth_val = validate_an_epoch_graph(model, dataset_val, val_steps)
-        print(f"\n[Validation Epoch MEAN], result: loss={result_val[0]:1.4f}, "
+        result_val, depth_val = validater_graph.run_an_epoch(model, dataset_val, val_steps)
+        print(f"[Validation Epoch MEAN], result: loss={result_val[0]:1.4f}, "
               f"trj_err={result_val[1]:1.4f}, rot_err={result_val[2]:1.4f}")
 
-        # if epoch % 10 == 0:
-        #     print("save intermediate results ...")
-        #     log.save_reconstruction_samples(model, dataset_val, epoch)
-        #     log.save_loss_scales(model, dataset_val, val_steps, opts.STEREO)
-        # save_model(model, result_val[0])
+        if epoch % 10 == 0:
+            print("save intermediate results ...")
+            log.save_reconstruction_samples(model, dataset_val, epoch)
+            log.save_loss_scales(model, dataset_val, val_steps, opts.STEREO)
+        save_model(model, result_val[0])
         log.save_log(epoch, result_train, result_val, depth_train, depth_val)
 
 
@@ -107,171 +107,6 @@ def get_dataset(dataset_name, split, shuffle, batch_size=opts.BATCH_SIZE):
     dataset = TfrecordGenerator(tfr_train_path, shuffle=shuffle, batch_size=batch_size).get_generator()
     steps_per_epoch = uf.count_steps(tfr_train_path, batch_size)
     return dataset, steps_per_epoch
-
-
-# Eager training is slow ...
-def train_an_epoch_eager(model, dataset, optimizer, steps_per_epoch):
-    results = []
-    depths = []
-    compute_loss = loss_factory()
-    # tf.data.Dataset object is reusable after a full iteration, check test_reuse_dataset()
-    for step, features in enumerate(dataset):
-        start = time.time()
-        with tf.GradientTape() as tape:
-            preds = model(features['image'])
-            loss_batch, loss_by_type = compute_loss(preds, features)
-
-        grads = tape.gradient(loss_batch, model.trainable_weights())
-        optimizer.apply_gradients(zip(grads, model.trainable_weights()))
-        loss_mean = tf.reduce_mean(loss_batch).numpy()
-        loss_by_type = tf.reduce_mean(loss_by_type, axis=1)
-
-        batch_result, log_msg, mean_depths = merge_results(features, preds, loss_mean, loss_by_type)
-        uf.print_progress_status(f"\tTraining (eager) {step}/{steps_per_epoch} steps, {log_msg}, "
-                                 f"time={time.time() - start:1.4f} ...")
-        results.append(batch_result)
-        depths.append(mean_depths)
-
-    # mean_res: mean of [all losses, trj_err, rot_err, weighted losses from various loss types]
-    mean_res = np.array(results).mean(axis=0)
-    depths = np.concatenate(depths, axis=0)
-    return mean_res, depths
-
-
-# Graph training is faster than eager training TWO TIMES!!
-def train_an_epoch_graph(model, dataset, optimizer, steps_per_epoch):
-    results = []
-    depths = []
-    compute_loss = loss_factory()
-    # tf.data.Dataset object is reusable after a full iteration, check test_reuse_dataset()
-    for step, features in enumerate(dataset):
-        start = time.time()
-        preds, loss, loss_by_type = train_a_batch(model, features, optimizer, compute_loss)
-        batch_result, log_msg, mean_depths = merge_results(features, preds, loss, loss_by_type)
-        uf.print_progress_status(f"\tTraining (graph) {step}/{steps_per_epoch} steps, {log_msg}, "
-                                 f"time={time.time() - start:1.4f} {mean_depths.shape}...")
-        results.append(batch_result)
-        depths.append(mean_depths)
-
-    print("")
-    # mean_res: mean of [all losses, trj_err, rot_err, weighted losses from various loss types]
-    mean_res = np.array(results).mean(axis=0)
-    depths = np.concatenate(depths, axis=1)
-    stride = depths.shape[1] // 10
-    depths = depths[:, 0:-1:stride]
-    return mean_res, depths
-
-
-@tf.function
-def train_a_batch(model, features, optimizer, compute_loss):
-    with tf.GradientTape() as tape:
-        # NOTE! preds = {"disp_ms": ..., "pose": ...} = model(image)
-        preds = model(features)
-        loss_batch, loss_by_type = compute_loss(preds, features)
-
-    grads = tape.gradient(loss_batch, model.trainable_weights())
-    optimizer.apply_gradients(zip(grads, model.trainable_weights()))
-    loss_mean = tf.reduce_mean(loss_batch)
-    loss_by_type = tf.reduce_mean(loss_by_type, axis=1)
-    """
-    preds: {"pose": ..., "disp_ms":, ...}
-    loss_mean: loss scalar that is averaged over all this epoch  
-    loss_by_type: loss [loss types]
-    """
-    return preds, loss_mean, loss_by_type
-
-
-def validate_an_epoch_eager(model, dataset, steps_per_epoch):
-    results = []
-    depths = []
-    compute_loss = loss_factory()
-    # tf.data.Dataset 객체는 한번 쓴 후에도 다시 iteration 가능, test_reuse_dataset() 참조
-    for step, features in enumerate(dataset):
-        start = time.time()
-        preds = model(features)
-
-        loss_batch, loss_by_type = compute_loss(preds, features)
-        loss_mean = tf.reduce_mean(loss_batch).numpy()
-        loss_by_type = tf.reduce_mean(loss_by_type, axis=1)
-
-        batch_result, log_msg, mean_depths = merge_results(features, preds, loss_mean, loss_by_type)
-        uf.print_progress_status(f"\tValidating (eager) {step}/{steps_per_epoch} steps, {log_msg}, "
-                                 f"time={time.time() - start:1.4f} ...")
-        results.append(batch_result)
-        depths.append(mean_depths)
-
-    # mean_res: mean of [all losses, trj_err, rot_err, weighted losses from various loss types]
-    mean_res = np.array(results).mean(axis=0)
-    depths = np.concatenate(depths, axis=0)
-    return mean_res, depths
-
-
-def validate_an_epoch_graph(model, dataset, steps_per_epoch):
-    results = []
-    depths = []
-    compute_loss = loss_factory()
-    for step, features in enumerate(dataset):
-        start = time.time()
-        preds, loss, loss_by_type = validate_a_batch(model, features, compute_loss)
-        batch_result, log_msg, mean_depths = merge_results(features, preds, loss, loss_by_type)
-        uf.print_progress_status(f"\tValidating (graph) {step}/{steps_per_epoch} steps, {log_msg}, "
-                                 f"time={time.time() - start:1.4f} {mean_depths.shape}...")
-        results.append(batch_result)
-        depths.append(mean_depths)
-
-    print("")
-    # mean_res: mean of [all losses, trj_err, rot_err, weighted losses from various loss types]
-    mean_res = np.array(results).mean(axis=0)
-    depths = np.concatenate(depths, axis=1)
-    stride = depths.shape[1] // 10
-    depths = depths[:, 0:-1:stride]
-    return mean_res, depths
-
-
-@tf.function
-def validate_a_batch(model, features, compute_loss):
-    preds = model(features)
-    loss_batch, loss_by_type = compute_loss(preds, features)
-    loss_mean = tf.reduce_mean(loss_batch)
-    loss_by_type = tf.reduce_mean(loss_by_type, axis=1)
-    return preds, loss_mean, loss_by_type
-
-
-def merge_results(features, preds, loss, loss_by_type):
-    mean_depths = get_center_depths(features, preds)
-    trjerr, roterr = get_metric_pose(preds, features)
-    batch_result = [loss.numpy(), trjerr, roterr] + loss_by_type.numpy().tolist()
-    log_msg = f"loss = {loss.numpy():1.4f}, metric={trjerr:1.4f}, {roterr:1.4f}"
-    return batch_result, log_msg, mean_depths
-
-
-def get_center_depths(features, preds):
-    depth_true = features["depth_gt"].numpy()
-    pred_disp_ms = preds["disp_ms"]
-    pred_depth_ms = uf.disp_to_depth_tensor(pred_disp_ms)
-    depth_pred = pred_depth_ms[0].numpy()
-    batch, height, width, _ = depth_true.shape
-    xs, xe = width // 2 - 10, width // 2 + 10
-    ys, ye = height // 4 * 3 - 10, height // 4 * 3 + 10
-
-    depth_true = depth_true[:, ys:ye, xs:xe, :]
-    mean_true = []
-    for depth in depth_true:
-        mean_d = depth[depth > 0].mean()
-        mean_true.append(mean_d)
-    mean_true = np.array(mean_true)
-    depth_pred = depth_pred[:, ys:ye, xs:xe, :]
-    mean_pred = depth_pred.mean(axis=(1, 2, 3))
-    mean_depths = np.stack([mean_true, mean_pred], axis=0)
-    return mean_depths
-
-
-def get_metric_pose(preds, features):
-    if "pose_gt" in features:
-        trjerr, roterr = compute_metric_pose(preds['pose'], features['pose_gt'])
-        return trjerr.numpy(), roterr.numpy()
-    else:
-        return 0, 0
 
 
 def save_model(model, val_loss):
